pyproject.toml
[tool.poetry]
name = "django-samples"
version = "0.1.0"
description = ""
authors = ["이한영 <dev@lhy.kr>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
django-extra-fields = "^2.0.3"
djangorestframework = "^3.11.0"
django-extensions = "^2.2.8"
pillow = "^7.0.0"
pytest = "^5.4.1"
pytest-django = "^3.8.0"
drf-yasg = "^1.17.1"

[tool.poetry.dev-dependencies]
notebook = "^6.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
